requirements.txt
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-dotenv==1.0.0
argparse==1.4.0
tabulate==0.9.0
